requirements.txt
flask
flask-cors
langchain-text-splitters
langchain-community
beautifulsoup4
lxml
sentence-transformers
chromadb
scikit-learn
numpy
groq
python-dotenv
